Projet IA et Robotique/pyproject.toml
[project]
name = "projet-ia-et-robotique"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13"
dependencies = [
    "guardrails-ai>=0.6.7",
    "ollama>=0.6.1",
    "pysimplegui>=5.0.8.3",
    "questionary>=2.1.1",
    "rich>=14.2.0",
]
